math64.h: Add mul_s64_s64_shr()
This function is needed for the nested TSC scaling code because if we
don't treat the offset values as signed then the result overflows.

The new function is almost identical to the existing mul_u64_u64_shr()
so this patch extracts the common code into a new mul_u64_u64() helper.

Signed-off-by: Ilias Stamatis <ilstam@amazon.com>