KVM: Generalize coalesced_mmio_has_room()
The KVM MMIO coalescing API supports a single ring buffer shared by all
coalesced MMIO zones. As a result coalesced_mmio_has_room() and
coalesced_mmio_write() operate on that ring buffer directly.

Change this by passing a ring pointer along with a ring size to
coalesced_mmio_has_room() to make the interface more abstract and allow
it to operate on any ring.

This is preliminary work to allow the coalescing API to work with
different ring buffers of varying sizes. As such, tidy the code in
coalesced_mmio_write() a bit in a way that makes subsequent patches
easier to review.

Signed-off-by: Ilias Stamatis <ilstam@amazon.com>

CR: https://code.amazon.com/reviews/CR-135763697